The refactor before the storm
Use i64 on expr, for when we eventually support it. Extract number
parsing function. Clean up

+++ src/ast.rs
@@ -111,7 +111,7 @@ impl fmt::Display for FuncDef {
 
 #[derive(Debug, PartialEq, Eq)]
 pub enum Expr {
-    Literal(u8),
+    Literal(i64),
     Ident(String),
 }
 impl Expr {
